ifstatus: show a no-carrier status (bsc#=1160654)
Split setup-in-progress and show a (transient) no-carrier status
when the device is set administratively up (device-up), but the
(L2) link is not up and running yet (link-up), including e.g.
not-associated wireless links.
The follwing (L3) setup like IP / lease acquisition continues to
appear as setup-in-progress until the setup is sufficient and we
consider and show it as up.